requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
brotli=1.0.9=h0d85af4_6
brotli-bin=1.0.9=h0d85af4_6
brotlipy=0.7.0=py37h271585c_1003
ca-certificates=2021.10.8=h033912b_0
certifi=2021.10.8=py37hf985489_1
cffi=1.15.0=py37h446072c_0
charset-normalizer=2.0.9=pyhd8ed1ab_0
cryptography=36.0.0=py37h5e77fcc_0
cycler=0.11.0=pyhd8ed1ab_0
fonttools=4.28.3=py37h271585c_0
freetype=2.10.4=h4cff582_1
idna=3.1=pyhd3deb0d_0
jbig=2.1=h0d85af4_2003
jpeg=9d=hbcb3906_0
kiwisolver=1.3.2=py37h737db71_1
lcms2=2.12=h577c468_0
lerc=3.0=he49afe7_0
libblas=3.9.0=12_osx64_openblas
libbrotlicommon=1.0.9=h0d85af4_6
libbrotlidec=1.0.9=h0d85af4_6
libbrotlienc=1.0.9=h0d85af4_6
libcblas=3.9.0=12_osx64_openblas
libcxx=12.0.1=habf9029_0
libdeflate=1.8=h0d85af4_0
libffi=3.4.2=h0d85af4_5
libgfortran=5.0.0=9_3_0_h6c81a4c_23
libgfortran5=9.3.0=h6c81a4c_23
liblapack=3.9.0=12_osx64_openblas
libopenblas=0.3.18=openmp_h3351f45_0
libpng=1.6.37=h7cec526_2
libtiff=4.3.0=hd146c10_2
libwebp-base=1.2.1=h0d85af4_0
libzlib=1.2.11=h9173be1_1013
llvm-openmp=12.0.1=hda6cdc1_1
lz4-c=1.9.3=he49afe7_1
matplotlib=3.5.1=py37hf985489_0
matplotlib-base=3.5.1=py37h3147e9e_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.2=h2e338ed_4
numpy=1.21.4=py37haefe36b_0
olefile=0.46=pyh9f0ad1d_1
openjdk=11.0.9.1=hcf210ce_1
openjpeg=2.4.0=h6e7aa92_1
openssl=1.1.1l=h0d85af4_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.3.5=py37h5b83a90_0
pillow=8.4.0=py37h76dc067_0
pip=21.3.1=pyhd8ed1ab_0
pycparser=2.21=pyhd8ed1ab_0
pyopenssl=21.0.0=pyhd8ed1ab_0
pyparsing=3.0.6=pyhd8ed1ab_0
pypdf2=1.26.0=py_2
pysocks=1.7.1=py37hf985489_4
python=3.7.12=haf480d7_100_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.7=2_cp37m
pytz=2021.3=pyhd8ed1ab_0
readline=8.1=h05e3726_0
requests=2.26.0=pyhd8ed1ab_1
selenium=3.141.0=py37h271585c_1003
setuptools=59.6.0=py37hf985489_0
setuptools-scm=6.3.2=pyhd8ed1ab_0
setuptools_scm=6.3.2=hd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.37.0=h23a322b_0
tabula-py=2.2.0=py37hf985489_1
tk=8.6.11=h5dbffcc_1
tomli=2.0.0=pyhd8ed1ab_1
tornado=6.1=py37h271585c_2
unicodedata2=13.0.0.post2=py37h271585c_4
urllib3=1.26.7=pyhd8ed1ab_0
wheel=0.37.0=pyhd8ed1ab_1
xz=5.2.5=haf1e3a3_1
zlib=1.2.11=h9173be1_1013
zstd=1.5.0=h582d3a0_0
